move funcs for meta to a separate .go file
